requirements.txt
typing_extensions
torch
torchaudio
torchvision
yolov5
Pillow